feat: add custom 404 page to avoid Next.js warning
