+++ components/Header/Header.scss
@@ -183,6 +183,7 @@
   /*** 
     Login - no searchbar, only feedback nav item 
   ***/
+  &--404,
   &--login,
   &--register,
   &--confirm,
@@ -197,6 +198,7 @@
   /***  
     Blue bG bar 
   ***/
+  &--404,
   &--results,
   &--video,
   &--article,
