+++ pages/404.js
@@ -1,0 +1,5 @@
+import ErrorSection from 'components/errors/ErrorSection';
+
+const FourOhFour = () => <ErrorSection statusCode={ 404 } title="404 - Page Not Found" />;
+
+export default FourOhFour;
